Setup the TRON Game and updated the SUBRISE.js library to v1.0.1
The Game class has been tweaked a bit, when you initialize a game you can set the
width and height of the game canvas. The window size is default. After you've intialized
the game, the game runs itself.

New classes are:
- GameComponent
- Point
- Rectangle

A GameComponent is an object with an update and draw method. When the GameComponent is
added to a game. It's update and draw method are automatically called.

Point is a simple object with an x and y property.

Rectangle helps to create virtual rectangles on a 2d matrix. The rectangle should hold an
easy collision detection method. Should as in, I haven't tested it but in theory it SHOULD work.
The rectangle class also caches certain values.

Setup TRON Game ready to start programming and using the new Subrise JavaScript Library. Kept the
Subrise Library seperate, but will need to concatenate it into custom.js before sending it in.