tweaked roll over settings, and made it fullscreen
made the game fullscreen since it's a techdemo and
fullscreen makes it more fun. Also tweaked some
parts where it felt the plane was twitching.
Something that only occured when the plane had
more space to fly. So fullscreen might be a
better choice to develop for now. It also uses
more resources I think, so that I can optimize
sooner.